requirements.txt
streamlit>=1.33
openpyxl>=3.1
httpx>=0.27
selectolax>=0.3.21
beautifulsoup4>=4.12
python-dotenv>=1.0
pyyaml>=6.0
pandas>=2.0
